backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pyjwt
cryptography
python-multipart
# click-8.3.0 h11-0.16.0 psycopg2-binary-2.9.11 sqlalchemy-2.0.44 uvicorn-0.38.0